aiad/src/requirements.txt
influxdb-client==1.45.0
jproperties==2.1.2
matplotlib
numpy
pandas==2.2.2
python-dateutil==2.8.2
requests==2.32.3
scikit-learn==1.4.2
tensorflow==2.16.1
python-qpid-proton==0.34.0
